dom/devicestorage/test/test_dotdot.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html> <!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717103
-->
<head>
  <title>Test for the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717103">Mozilla Bug 717103</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

devicestorage_setup(function () {

function testingStorage() {
  return navigator.getDeviceStorage("pictures");
}

var tests = [
  function () { return testingStorage().addNamed(createRandomBlob('image/png'), gFileName); },
  function () { return testingStorage().delete(gFileName); },
  function () { return testingStorage().get(gFileName); },
  function () { var r = testingStorage().enumerate("../"); return r; }
];

var gFileName = "../owned.png";

function fail(e) {
  ok(false, "addSuccess was called");
  dump(request);
  SimpleTest.finish();
}

function next(e) {

  if (e != undefined) {
    ok(true, "addError was called");
    ok(e.target.error.name == "SecurityError", "Error must be SecurityError");
  }

  var f = tests.pop();

  if (f == undefined) {
    SimpleTest.finish();
    return;
  }

  request = f();
  request.onsuccess = fail;
  request.onerror = next;
}

next();

});

</script>
</pre>
</body>
</html>
